+++ packages/gatsby-source-medium/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.4.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-medium@2.4.0-next.0...gatsby-source-medium@2.4.0-next.1) (2020-11-10)
+
+**Note:** Version bump only for package gatsby-source-medium
+
 ## [2.3.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-medium@2.3.13...gatsby-source-medium@2.3.14) (2020-10-26)
 
 ### Bug Fixes